Implemented Compile trait for IfStatement
Implementation requires compiler to keep track of labels, and contexts.

Labels are created by the compiler with/without target. Compiler
maintains the indexes of instructions that refer to a label, and when
the label is placed relevant instructions are updated. Removing labels
from the compiler's map is the responsibility of the programmer.